requirement.txt
absl-py==0.12.0
affine==2.3.0
astor==0.8.1
attrs==21.2.0
backcall==0.2.0
cached-property==1.5.2
certifi==2020.12.5
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
contextlib2==0.6.0.post1
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==5.0.3
descartes==1.1.0
Fiona==1.8.19
Flask==2.0.1
gast==0.2.2
geopandas==0.9.0
google-pasta==0.2.0
grpcio==1.36.1
h5py==3.1.0
imageio==2.9.0
importlib-metadata==3.10.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lvis==0.5.3
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.3.4
munch==2.5.0
networkx==2.5
numpy==1.19.5
opencv-contrib-python==4.5.2.52
opencv-python==4.5.1.48
opencv-python-headless==4.5.2.52
opt-einsum==3.3.0
pandas==1.1.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prompt-toolkit==3.0.18
protobuf==3.15.7
ptyprocess==0.7.0
pycocotools==2.0.2
Pygments==2.9.0
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
rasterio==1.2.3
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
snuggs==1.4.7
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
tf-slim==1.1.0
tifffile==2020.9.3
tk==0.1.0
tqdm==4.59.0
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.1
